+++ src/libs/src/dijkstra.c
@@ -5,7 +5,7 @@
 #include "../libs/PQ.h"
 #include "../libs/dijkstra.h"
 
-void initialize_single_source(Graph *graph, PQ *h, double *dist, int src) {
+void initialize_single_source(Graph *graph, PQ *h, long double *dist, int src) {
     int num_nodes = graph_get_num_nodes(graph);
     for (int i = 0; i < num_nodes; i++) {
         PQ_insert(h, i, i == src ? 0.0 : DBL_MAX);
@@ -13,12 +13,12 @@ void initialize_single_source(Graph *graph, PQ *h, double *dist, int src) {
     }
 }
 
-void relax(Graph *graph, PQ *h, double *dist, int u) {
+void relax(Graph *graph, PQ *h, long double *dist, int u) {
     int num_nodes = graph_get_num_nodes(graph);
-    double u_priority = PQ_get_priority(h, u);
+    long double u_priority = PQ_get_priority(h, u);
 
     for (int i = 0; i < num_nodes; i++) {
-        double w = graph_get_cost_from_edge(graph, u, i);
+        long double w = graph_get_cost_from_edge(graph, u, i);
         if (dist[i] > u_priority + w && w != -1) {
             dist[i] = u_priority + w;
             PQ_decrease_key(h, i, u_priority + w);
@@ -26,10 +26,10 @@ void relax(Graph *graph, PQ *h, double *dist, int u) {
     }
 }
 
-double *dijkstra_algorithm(Graph *graph, int src) {
+long double *dijkstra_algorithm(Graph *graph, int src) {
     int num_nodes = graph_get_num_nodes(graph);
     PQ *h = PQ_init(num_nodes);
-    double *dist = calloc(num_nodes, sizeof(double));
+    long double *dist = calloc(num_nodes, sizeof(long double));
 
     initialize_single_source(graph, h, dist, src);
